third_party/WebKit/LayoutTests/animations/computed-style-expected.txt

This tests computed style values from animation properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test1Style.animationName is 'anim1'
PASS test2Style.animationName is 'anim2, anim3'
PASS test3Style.animationName is 'anim1, anim2, anim3'
PASS test1Style.animationDuration is '10s'
PASS test2Style.animationDuration is '5s, 2.5s'
PASS test1Style.animationFillMode is 'backwards'
PASS test2Style.animationFillMode is 'forwards, both'
PASS test1Style.animationIterationCount is '10'
PASS test2Style.animationIterationCount is '10, infinite'
PASS test1Style.animationTimingFunction is 'linear'
PASS test2Style.animationTimingFunction is 'linear, ease-in-out'
PASS test3Style.animationTimingFunction is 'step-start, step-end, steps(5, end)'
PASS test1Style.animationDirection is 'normal'
PASS test2Style.animationDirection is 'normal, alternate'


PASS successfullyParsed is true

TEST COMPLETE
